gadget/install: fix nosecboot unit tests
